build_targets/tf_keras_pillow/requirements.txt
tensorflow==1.8.0
keras
pillow
